Locking Down DeleteItem Permissions

+++ backend/src/resolvers/Mutation.js
@@ -51,9 +51,17 @@ const Mutations = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
     // 1. find the item
-    const item = await ctx.db.query.item({ where }, `{ id title}`);
+    const item = await ctx.db.query.item({ where }, `{ id title user { id }}`);
     // 2. Check if they own that item, or have the permissions
-    // TODO
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ['ADMIN', 'ITEMDELETE'].includes(permission)
+    );
+
+    if (!ownsItem && !hasPermissions) {
+      throw new Error("You don't have permission to do that!");
+    }
+
     // 3. Delete it!
     return ctx.db.mutation.deleteItem({ where }, info);
   },
@@ -130,16 +138,13 @@ const Mutations = {
 
     // 3. Email them that reset token
     const mailRes = await transport.sendMail({
-      from: 'saglam.chd@gmail.com',
+      from: 'wes@wesbos.com',
       to: user.email,
       subject: 'Your Password Reset Token',
       html: makeANiceEmail(`Your Password Reset Token is here!
-      \n
-      \n
-      <a href="${process.env.FRONTEND_URL}/reset?resetToken=${resetToken}">
-        Click Here to Reset
-      </a>
-      `),
+      \n\n
+      <a href="${process.env
+        .FRONTEND_URL}/reset?resetToken=${resetToken}">Click Here to Reset</a>`),
     });
 
     // 4. Return the message
@@ -176,7 +181,7 @@ const Mutations = {
         resetTokenExpiry: null,
       },
     });
-    
+
     // 6. Generate JWT
     const token = jwt.sign({ userId: updatedUser.id }, process.env.APP_SECRET);
     
@@ -189,7 +194,7 @@ const Mutations = {
     // 8. return the new user
     return updatedUser;
   },
-
+  
   async updatePermissions(parent, args, ctx, info) {
     // 1. Check if they are logged in
     if (!ctx.request.userId) {
